feature(js): lightbox is now bootstrapped at runtime
Lightbox is now bootstrapped at runtime. Colorbox however is not required until
elgg/lightbox#open is called. Lightbox css now extends elgg.css.
